+++ packages/core/BaseFeatureWidget/BaseFeatureDetail/types.ts
@@ -1,0 +1,8 @@
+import type { BaseCardProps } from '../types'
+
+export interface BaseInputProps extends BaseCardProps {
+  omit?: string[]
+  model: any
+  descriptions?: Record<string, React.ReactNode>
+  formatter?: (val: unknown, key: string) => React.ReactNode
+}
